Send request with JSON data and encode image as base64 string
Previously, the POST request sent a formdata object containing the image file and metadata.
This resulted in many issues due to AWS Lambda treating the formdata object as a raw string, with no easy way to extract its keys/values.

Now, the request simply sends a JSON object which represents the image as a base64-encoded string.
This assumes that the backend will be able to decode base64.